Create Sample 1
Notice that when the game is started over for the 3rd time, the computer started copying the last move made by the player. For example, when the player chose rock in the 2nd round, the computer copied the player's move and hence chose rock in the 3rd round. The player exploited this known fact so as to win the game.